Move unit tests to memory
