+++ src/vm/memory.rs
@@ -10,3 +10,17 @@ impl Memory {
         return self.data.get(&addr);
     }
 }
+
+#[cfg(test)]
+pub mod memory_tests {
+    #[test]
+    pub fn get_test () {
+        let key = MaybeRelocatable::Int(BigInt::from(2))
+        let val = MaybeRelocatable::Int(BigInt::from(5))
+        let mem:Memory {
+            HashMap::from([(key, val)])
+        };
+        assert_eq!(mem.get(&key), MaybeRelocatable::Int(BigInt::from(5)));
+    }
+}
+
